💄 create story and design ProfileRadioGroup
